requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
arrow==1.2.3
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrdict==2.0.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blessed==1.20.0
blis==0.7.9
boltons==23.0.0
brotlipy==0.7.0
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.4
cloudpickle==2.2.1
cmake==3.26.4
colorama==0.4.4
comm==0.1.2
conda==23.5.0
conda-content-trust==0+unknown
conda-package-handling==2.1.0
conda_package_streaming==0.8.0
confection==0.1.0
contourpy==1.1.0
croniter==1.3.15
cryptography==38.0.4
cycler==0.11.0
cymem==2.0.7
dask==2023.7.0
dateutils==0.6.12
debugpy==1.5.1
decorator==5.1.1
deepdiff==6.3.1
dill==0.3.4
entrypoints==0.4
exceptiongroup==1.1.2
executing==0.8.3
fastapi==0.100.0
filelock==3.12.2
flatbuffers==2.0.7
fonttools==4.42.1
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.4.0
gensim==4.2.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.51.3
h11==0.14.0
h5py==3.9.0
huggingface-hub==0.16.4
idna==3.3
importlib-metadata==6.0.0
importlib-resources==6.0.1
inquirer==3.1.3
IProgress==0.4
ipykernel==6.19.2
ipython==8.12.0
ipywidgets==8.0.7
itsdangerous==2.1.2
jedi==0.18.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.1
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.8
keras==2.7.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.1.1
kiwisolver==1.4.5
langcodes==3.3.0
libclang==16.0.6
lightning==2.0.4
lightning-cloud==0.5.37
lightning-utilities==0.9.0
lit==16.0.6
locket==1.0.0
loguru==0.7.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.535.133
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvitop==1.3.2
oauthlib==3.2.2
opt-einsum==3.3.0
ordered-set==4.1.0
packaging==21.3
pandarallel==1.6.5
pandas==2.0.3
parso==0.8.3
partd==1.4.0
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip==23.1.2
platformdirs==2.5.2
pluggy==1.0.0
preshed==3.0.8
pretty-errors==1.2.25
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycosat==0.6.3
pycparser==2.21
pydantic==1.10.11
Pygments==2.15.1
PyJWT==2.7.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.4
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
ray==2.5.1
readchar==4.0.5
referencing==0.30.0
regex==2023.6.3
requests==2.29.0
requests-oauthlib==1.3.1
rich==13.4.2
rpds-py==0.9.2
rsa==4.9
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.6
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.10.1
seqeval==1.2.2
setuptools==67.8.0
shortuuid==1.0.13
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.4.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.2.0
starlette==0.27.0
starsessions==1.3.0
swifter==1.3.5
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.7.4
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.4
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
thinc==8.1.10
threadpoolctl==3.1.0
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torch-tb-profiler==0.4.1
torchmetrics==1.0.0
tornado==6.2
tqdm==4.63.0
traitlets==5.9.0
transformers==4.30.2
triton==2.0.0
typer==0.7.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.8
uvicorn==0.22.0
wasabi==0.10.1
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.3.6
wheel==0.38.4
widgetsnbextension==4.0.8
wrapt==1.15.0
yarl==1.9.2
zipp==3.11.0
zstandard==0.15.2
